+++ src/plugins/rm/rm.c
@@ -208,7 +208,7 @@ _parse_cont_header(int fd, struct rm_info *info)
 static void *
 _match(struct plugin *p, const char *path, int len, int base)
 {
-    int i;
+    long i;
 
     i = lms_which_extension(path, len, _exts, LMS_ARRAY_SIZE(_exts));
     if (i < 0)
@@ -264,7 +264,7 @@ _parse(struct plugin *plugin, struct lms_context *ctxt, const struct lms_file_in
 
     /* try to define stream type by extension */
     if (stream_type == STREAM_TYPE_UNKNOWN) {
-        int ext_idx = ((int)match) - 1;
+        long ext_idx = ((long)match) - 1;
         if (strcmp(_exts[ext_idx].str, ".ra") == 0)
             stream_type = STREAM_TYPE_AUDIO;
         /* consider rv, rm, rmj and rmvb as video */
@@ -276,8 +276,8 @@ _parse(struct plugin *plugin, struct lms_context *ctxt, const struct lms_file_in
     lms_string_size_strip_and_free(&info.artist);
 
     if (!info.title.str) {
-        int ext_idx;
-        ext_idx = ((int)match) - 1;
+        long ext_idx;
+        ext_idx = ((long)match) - 1;
         info.title.len = finfo->path_len - finfo->base - _exts[ext_idx].len;
         info.title.str = malloc((info.title.len + 1) * sizeof(char));
         memcpy(info.title.str, finfo->path + finfo->base, info.title.len);
